Minor code improvement
I'm slowly getting a hang of TornadoFX's DSL-like hilarity. Which
frankly I'd get a hold of faster if I dedicated myself to reading the
docs and going through them but that's not how I roll in my off-time,
I'm here to hack blindly and make dumb commit messages, and I'm all out
of...dumb...commit...messages? well that's manifestly false