Hacked in a sync elevation button
ok the color thing is...not great. but ok. it's fine. it's fiiiine. i
mean.

look I could either hack on this or watch The Lighthouse and eat dinner
with friends and I choose the Social Thing. hence why there be a pair of
janky maps splatted in the center of the file

and why MainScreen is a utter, total mess. rip.